Remove inconsistencies in scipy.misc / scipy.ndimage
* Move scipy.misc.imfilter to scipy.ndimage.imfilter
* Move scipy.misc.imresize to scipy.ndimage.imresize
* Move scipy.misc.imrotate to scipy.ndimage.imrotate

Fixes #6242
