+++ scipy/misc/pilutil.py
@@ -12,14 +12,13 @@ from __future__ import division, print_function, absolute_import
 import numpy
 import tempfile
 
-from numpy import (amin, amax, ravel, asarray, cast, arange, ones, newaxis,
-                   transpose, iscomplexobj, uint8, issubdtype, array)
+from numpy import (amin, amax, ravel, asarray, arange, ones, newaxis,
+                   transpose, iscomplexobj, uint8, array)
 
 try:
-    from PIL import Image, ImageFilter
+    from PIL import Image
 except ImportError:
     import Image
-    import ImageFilter
 
 
 if not hasattr(Image, 'frombytes'):
@@ -27,6 +26,20 @@ if not hasattr(Image, 'frombytes'):
 
 __all__ = ['fromimage', 'toimage', 'imsave', 'imread', 'bytescale',
            'imrotate', 'imresize', 'imshow', 'imfilter']
+
+from scipy.ndimage import imresize, imfilter, imrotate
+imresize = numpy.deprecate(imresize,
+                           message=('misc.imresize is deprecated in '
+                                    'Scipy 1.0. Please use '
+                                    'scipy.misc.ndimage.imresize instead'))
+imfilter = numpy.deprecate(imfilter,
+                           message=('misc.imfilter is deprecated in '
+                                    'Scipy 1.0. Please use '
+                                    'scipy.misc.ndimage.imfilter instead'))
+imrotate = numpy.deprecate(imrotate,
+                           message=('misc.imrotate is deprecated in '
+                                    'Scipy 1.0. Please use '
+                                    'scipy.misc.ndimage.imrotate instead'))
 
 
 # Returns a byte-scaled image
@@ -381,40 +394,6 @@ def toimage(arr, high=255, low=0, cmin=None, cmax=None, pal=None,
     return image
 
 
-def imrotate(arr, angle, interp='bilinear'):
-    """
-    Rotate an image counter-clockwise by angle degrees.
-
-    This function is only available if Python Imaging Library (PIL) is installed.
-
-    Parameters
-    ----------
-    arr : ndarray
-        Input array of image to be rotated.
-    angle : float
-        The angle of rotation.
-    interp : str, optional
-        Interpolation
-
-        - 'nearest' :  for nearest neighbor
-        - 'bilinear' : for bilinear
-        - 'lanczos' : for lanczos
-        - 'cubic' : for bicubic
-        - 'bicubic' : for bicubic
-
-    Returns
-    -------
-    imrotate : ndarray
-        The rotated array of image.
-
-    """
-    arr = asarray(arr)
-    func = {'nearest': 0, 'lanczos': 1, 'bilinear': 2, 'bicubic': 3, 'cubic': 3}
-    im = toimage(arr)
-    im = im.rotate(angle, resample=func[interp])
-    return fromimage(im)
-
-
 def imshow(arr):
     """
     Simple showing of an image through an external viewer.
@@ -457,96 +436,3 @@ def imshow(arr):
     os.unlink(fname)
     if status != 0:
         raise RuntimeError('Could not execute image viewer.')
-
-
-def imresize(arr, size, interp='bilinear', mode=None):
-    """
-    Resize an image.
-
-    This function is only available if Python Imaging Library (PIL) is installed.
-
-    Parameters
-    ----------
-    arr : ndarray
-        The array of image to be resized.
-
-    size : int, float or tuple
-        * int   - Percentage of current size.
-        * float - Fraction of current size.
-        * tuple - Size of the output image.
-
-    interp : str, optional
-        Interpolation to use for re-sizing ('nearest', 'lanczos', 'bilinear', 'bicubic'
-        or 'cubic').
-
-    mode : str, optional
-        The PIL image mode ('P', 'L', etc.) to convert `arr` before resizing.
-
-    Returns
-    -------
-    imresize : ndarray
-        The resized array of image.
-
-    See Also
-    --------
-    toimage : Implicitly used to convert `arr` according to `mode`.
-    scipy.ndimage.zoom : More generic implementation that does not use PIL.
-
-    """
-    im = toimage(arr, mode=mode)
-    ts = type(size)
-    if issubdtype(ts, int):
-        percent = size / 100.0
-        size = tuple((array(im.size)*percent).astype(int))
-    elif issubdtype(type(size), float):
-        size = tuple((array(im.size)*size).astype(int))
-    else:
-        size = (size[1], size[0])
-    func = {'nearest': 0, 'lanczos': 1, 'bilinear': 2, 'bicubic': 3, 'cubic': 3}
-    imnew = im.resize(size, resample=func[interp])
-    return fromimage(imnew)
-
-
-def imfilter(arr, ftype):
-    """
-    Simple filtering of an image.
-
-    This function is only available if Python Imaging Library (PIL) is installed.
-
-    Parameters
-    ----------
-    arr : ndarray
-        The array of Image in which the filter is to be applied.
-    ftype : str
-        The filter that has to be applied. Legal values are:
-        'blur', 'contour', 'detail', 'edge_enhance', 'edge_enhance_more',
-        'emboss', 'find_edges', 'smooth', 'smooth_more', 'sharpen'.
-
-    Returns
-    -------
-    imfilter : ndarray
-        The array with filter applied.
-
-    Raises
-    ------
-    ValueError
-        *Unknown filter type.*  If the filter you are trying
-        to apply is unsupported.
-
-    """
-    _tdict = {'blur': ImageFilter.BLUR,
-              'contour': ImageFilter.CONTOUR,
-              'detail': ImageFilter.DETAIL,
-              'edge_enhance': ImageFilter.EDGE_ENHANCE,
-              'edge_enhance_more': ImageFilter.EDGE_ENHANCE_MORE,
-              'emboss': ImageFilter.EMBOSS,
-              'find_edges': ImageFilter.FIND_EDGES,
-              'smooth': ImageFilter.SMOOTH,
-              'smooth_more': ImageFilter.SMOOTH_MORE,
-              'sharpen': ImageFilter.SHARPEN
-              }
-
-    im = toimage(arr)
-    if ftype not in _tdict:
-        raise ValueError("Unknown filter type.")
-    return fromimage(im.filter(_tdict[ftype]))
